+++ src/data/v2/mod.rs
@@ -3,6 +3,8 @@
 
 /// Definitions for retrieval of market data bars.
 pub mod bars;
+/// Functionality for retrieval of most recent quotes.
+pub mod last_quote;
 
 // TODO: Remove this alias with the next compatibility breaking release.
 #[deprecated(note = "renamed to 'bars'; use that instead")]