+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Unreleased
 ----------
+- Introduced support for retrieving the last quote for a symbol
 - Aliased `data::v2::bars` module to `data::v2::stocks`
   - Deprecated `data::v2::stocks` module alias